+++ build.rs
@@ -134,5 +134,12 @@ fn main() -> Result<()> {
                 .compile("suppressSourceFortificationCheck");
         }
     }
+
+    if target_os == "android" {
+        let pkg_name = env!("CARGO_PKG_NAME");
+        let soname = format!("lib{}.so", pkg_name);
+        println!("cargo:rustc-cdylib-link-arg=-Wl,-soname,{}", soname);
+    }
+
     Ok(())
 }