+++ clis/tcli/src/main.rs
@@ -7,9 +7,10 @@ mod nord;
 use anyhow::{anyhow, Result};
 use clap::Parser;
 use dirs::home_dir;
+use parking_lot::Mutex;
 use regex::Regex;
-use std::io::Write;
-use telio_model::{api_config::Features, event::Event as DevEvent};
+use std::{io::Write, sync::Arc};
+use telio_model::{api_config::Features, config::Server, event::Event as DevEvent};
 
 #[derive(Parser)]
 struct Args {
@@ -31,7 +32,9 @@ fn main() -> Result<()> {
         .transpose()?
         .unwrap_or_default();
 
-    let mut cli = cli::Cli::new(features, token)?;
+    let derp_server = Arc::new(Mutex::<Option<Server>>::new(None));
+
+    let mut cli = cli::Cli::new(features, token, derp_server.clone())?;
     let mut stdout = std::io::stdout();
 
     let less_spam = args.less_spam;
@@ -83,17 +86,17 @@ fn main() -> Result<()> {
                 Event(e) => match *e {
                     DevEvent::Node { body: Some(b) } => {
                         println!(
-                            "event node: {:?}:{};  Path = {:?}",
-                            b.state.ok_or_else(|| anyhow!("empty state"))?,
-                            b.public_key,
-                            b.path
-                        );
-                    }
-                    DevEvent::Relay { body: Some(b) } => {
-                        println!(
-                            "event relay: {}",
+                            "event node: {:?}",
                             serde_json::to_string(&b).unwrap_or_else(|_| "".to_string())
                         );
+                    }
+                    DevEvent::Relay { body } => {
+                        if let Some(b) = body.as_ref() {
+                            println!(
+                                "event relay: {}",
+                                serde_json::to_string(&b).unwrap_or_else(|_| "".to_string())
+                            );
+                        }
                     }
                     _ => (),
                 },